Consolidate config schemas from 9 files to 3, delete mixin classes
- Merge data_pipeline.py + runtime.py → data.py (extraction, validation,
  transformation, Neo4j, DuckDB, paths, logging, metrics, CLI, categorization)
- Merge enrichment.py + fiscal.py + ml.py + reporting.py → domain.py
  (enrichment services, fiscal analysis, ML/PaECTER, reporting)
- Keep pipeline.py as root composition (PipelineConfig, PipelineMetadata)
- Delete common.py: replace PercentageMappingMixin/FloatRangeValidatorMixin
  with inline helper function and Pydantic Field(ge=, le=) constraints
- Update all direct imports (fiscal → domain) across enrichers, transformers, tests

Before: 9 files + __init__.py = 1,342 lines
After: 3 files + __init__.py = 1,196 lines, 7 fewer files to navigate
